refactor: make svg imports in icon examples vite-friendly
