+++ frontend/demo/component/icons/icons-color.ts
@@ -4,7 +4,7 @@ import '@vaadin/icon';
 import { html, LitElement } from 'lit';
 import { customElement } from 'lit/decorators.js';
 import { applyTheme } from 'Frontend/generated/theme';
-import codeBranch from '../../../../src/main/resources/icons/code-branch.svg';
+import codeBranch from '../../../../src/main/resources/icons/code-branch.svg?url';
 
 @customElement('icons-color')
 export class Example extends LitElement {
