+++ frontend/demo/component/icons/react/svg-standalone.tsx
@@ -1,7 +1,7 @@
 import { reactExample } from 'Frontend/demo/react-example'; // hidden-source-line
 import React from 'react';
 import { Icon } from '@vaadin/react-components/Icon.js';
-import codeBranchIcon from '../../../../../src/main/resources/icons/code-branch.svg';
+import codeBranchIcon from '../../../../../src/main/resources/icons/code-branch.svg?url';
 
 function Example() {
   return (
