+++ frontend/demo/component/icons/svg-sprites.ts
@@ -4,7 +4,7 @@ import '@vaadin/icon';
 import { html, LitElement } from 'lit';
 import { customElement } from 'lit/decorators.js';
 import { applyTheme } from 'Frontend/generated/theme';
-import solidSprite from '../../../../src/main/resources/icons/solid.svg';
+import solidSprite from '../../../../src/main/resources/icons/solid.svg?url';
 
 @customElement('svg-sprites')
 export class Example extends LitElement {
